Move FailureReason to common package. Rename TreasuryFailureReason to TreasuryEconomyFailureReason.
