+++ api/src/main/java/me/lokka30/treasury/api/common/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * This package holds API classes that are meant to be utilized across multiple APIs
+ *
+ * @author creatorfromhell
+ * @since {@link me.lokka30.treasury.api.economy.misc.EconomyAPIVersion#v1_0 v1.0}
+ */
+package me.lokka30.treasury.api.common;
