+++ api/src/main/java/me/lokka30/treasury/api/common/response/FailureReason.java
@@ -1,4 +1,4 @@
-package me.lokka30.treasury.api.economy.response;
+package me.lokka30.treasury.api.common.response;
 
 import org.jetbrains.annotations.NotNull;
